+++ src/lib.rs
@@ -1,14 +1,17 @@
-pub mod siena;
-pub mod yaml;
 pub mod frontmatter;
 pub mod providers;
+pub mod siena;
 pub mod utils;
+pub mod yaml;
 
 #[cfg(test)]
 mod siena_tests;
 
 #[cfg(test)]
 mod frontmatter_tests;
+
+#[cfg(test)]
+mod utils_tests;
 
 #[cfg(test)]
 mod yaml_tests;