+++ src/siena_tests.rs
@@ -1,72 +1,81 @@
+use crate::providers::local::LocalProvider;
+use crate::siena::{siena, Record, RecordSortOrder};
 use std::{collections::HashMap, env};
-use crate::siena::{Siena, RecordSortOrder, Store, Record};
 
-#[test]
-fn sort_test() {
-    let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
+fn record_1() -> Record {
+    Record {
         id: String::from("test"),
         collection: String::from("demo"),
         file_name: String::from("test.yaml"),
         data: HashMap::from([
             (String::from("title"), String::from("Bye, world")),
             (String::from("date"), String::from("2022-09-10")),
-        ])
-    };
+        ]),
+    }
+}
 
-    let expected_data_item_2 = Record {
+fn record_2() -> Record {
+    Record {
         id: String::from("test2"),
         collection: String::from("demo"),
         file_name: String::from("test2.yml"),
         data: HashMap::from([
             (String::from("title"), String::from("Hello, world")),
             (String::from("date"), String::from("2022-09-09")),
-        ])
-    };
+        ]),
+    }
+}
 
-    let expected_data_item_3 = Record {
+fn record_3() -> Record {
+    Record {
         id: String::from("2"),
         collection: String::from("demo"),
         file_name: String::from("2.yml"),
         data: HashMap::from([
             (String::from("title"), String::from("Bye, World")),
             (String::from("date"), String::from("2022-01-01")),
-        ])
-    };
+        ]),
+    }
+}
 
-    let expected_data_item_4 = Record {
+fn record_4() -> Record {
+    Record {
         id: String::from("1"),
         collection: String::from("demo"),
         file_name: String::from("1.yml"),
         data: HashMap::from([
             (String::from("title"), String::from("Hello, World")),
             (String::from("date"), String::from("2020-01-01")),
-        ])
-    };
+        ]),
+    }
+}
 
-    let expected_data_item_5 = Record {
+fn record_5() -> Record {
+    Record {
         id: String::from("3"),
         collection: String::from("demo"),
         file_name: String::from("3.yml"),
         data: HashMap::from([
             (String::from("special-item"), String::from("true")),
             (String::from("date"), String::from("1992-09-17")),
-        ])
+        ]),
+    }
+}
+
+#[test]
+fn sort_test() {
+    let root_dir = env::current_dir().unwrap();
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
-    
-    let expected = Vec::from([
-        expected_data_item, 
-        expected_data_item_2,
-        expected_data_item_3,
-        expected_data_item_4,
-        expected_data_item_5,
-    ]);
+    let store = siena(provider);
+
+    let expected = Vec::from([record_1(), record_2(), record_3(), record_4(), record_5()]);
 
     let result = store
         .collection("demo")
@@ -79,52 +88,36 @@ fn sort_test() {
 #[test]
 fn when_is_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("1"),
-        collection: String::from("demo"),
-        file_name: String::from("1.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Hello, World")),
-            (String::from("date"), String::from("2020-01-01")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
-
-    let expected = Vec::from([
-        expected_data_item, 
-    ]);
+    let store = siena(provider);
 
     let result = store
         .collection("demo")
         .when_is("title", "Hello, World")
         .get_all();
 
-    assert_eq!(result, expected);
+    assert_eq!(result, Vec::from([record_4()]));
 }
 
 #[test]
 fn when_isnt_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected = Record {
-        id: String::from("1"),
-        collection: String::from("demo"),
-        file_name: String::from("1.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Hello, World")),
-            (String::from("date"), String::from("2020-01-01")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
+    let store = siena(provider);
 
     let result = store
         .collection("demo")
@@ -136,95 +129,41 @@ fn when_isnt_test() {
         .get_first()
         .unwrap();
 
-    assert_eq!(result, expected);
+    assert_eq!(result, record_4());
 }
 
 #[test]
 fn when_has_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("3"),
-        collection: String::from("demo"),
-        file_name: String::from("3.yml"),
-        data: HashMap::from([
-            (String::from("special-item"), String::from("true")),
-            (String::from("date"), String::from("1992-09-17")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
+    let store = siena(provider);
 
-    let expected = Vec::from([
-        expected_data_item
-    ]);
+    let result = store.collection("demo").when_has("special-item").get_all();
 
-    let result = store
-        .collection("demo")
-        .when_has("special-item")
-        .get_all();
-
-    assert_eq!(result, expected);
+    assert_eq!(result, Vec::from([record_5()]));
 }
 
 #[test]
 fn when_hasnt_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("test"),
-        collection: String::from("demo"),
-        file_name: String::from("test.yaml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Bye, world")),
-            (String::from("date"), String::from("2022-09-10")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
+    let store = siena(provider);
 
-    let expected_data_item_2 = Record {
-        id: String::from("test2"),
-        collection: String::from("demo"),
-        file_name: String::from("test2.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Hello, world")),
-            (String::from("date"), String::from("2022-09-09")),
-        ])
-    };
-
-    let expected_data_item_3 = Record {
-        id: String::from("2"),
-        collection: String::from("demo"),
-        file_name: String::from("2.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Bye, World")),
-            (String::from("date"), String::from("2022-01-01")),
-        ])
-    };
-
-    let expected_data_item_4 = Record {
-        id: String::from("1"),
-        collection: String::from("demo"),
-        file_name: String::from("1.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Hello, World")),
-            (String::from("date"), String::from("2020-01-01")),
-        ])
-    };
-
-    let expected = Vec::from([
-        expected_data_item,
-        expected_data_item_2,
-        expected_data_item_3,
-        expected_data_item_4,
-    ]);
+    let expected = Vec::from([record_1(), record_2(), record_3(), record_4()]);
 
     let result = store
         .collection("demo")
@@ -238,56 +177,36 @@ fn when_hasnt_test() {
 #[test]
 fn when_matches_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("3"),
-        collection: String::from("demo"),
-        file_name: String::from("3.yml"),
-        data: HashMap::from([
-            (String::from("special-item"), String::from("true")),
-            (String::from("date"), String::from("1992-09-17")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
-
-    let expected = Vec::from([
-        expected_data_item
-    ]);
+    let store = siena(provider);
 
     let result = store
         .collection("demo")
         .when_matches("date", r"1992")
         .get_all();
 
-    assert_eq!(result, expected);
+    assert_eq!(result, Vec::from([record_5()]));
 }
 
 #[test]
 fn limit_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("test"),
-        collection: String::from("demo"),
-        file_name: String::from("test.yaml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Bye, world")),
-            (String::from("date"), String::from("2022-09-10")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
-
-    let expected = Vec::from([
-        expected_data_item
-    ]);
+    let store = siena(provider);
 
     let result = store
         .collection("demo")
@@ -295,32 +214,21 @@ fn limit_test() {
         .limit(1)
         .get_all();
 
-    assert_eq!(result, expected);
+    assert_eq!(result, Vec::from([record_1()]));
 }
 
 #[test]
 fn offset_test() {
     let root_dir = env::current_dir().unwrap();
-
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    let expected_data_item = Record {
-        id: String::from("test2"),
-        collection: String::from("demo"),
-        file_name: String::from("test2.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Hello, world")),
-            (String::from("date"), String::from("2022-09-09")),
-        ])
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
     };
-
-
-    let expected = Vec::from([
-        expected_data_item
-    ]);
+    let store = siena(provider);
 
     let result = store
         .collection("demo")
@@ -329,18 +237,22 @@ fn offset_test() {
         .limit(1)
         .get_all();
 
-    assert_eq!(result, expected);
+    assert_eq!(result, Vec::from([record_2()]));
 }
 
 #[test]
 fn offset_out_of_bounds_test() {
     let root_dir = env::current_dir().unwrap();
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
+    };
+    let store = siena(provider);
 
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-    
     let result = store
         .collection("demo")
         .sort("date", RecordSortOrder::Desc)
@@ -353,18 +265,21 @@ fn offset_out_of_bounds_test() {
 #[test]
 fn update_test() {
     let root_dir = env::current_dir().unwrap();
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
+    };
 
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-    
-    store
+    siena(provider.clone())
         .collection("demo")
         .when_is("date", "1992-09-17")
         .set(Vec::from([("special-item", "false")]));
 
-    let result = store
+    let result = siena(provider.clone())
         .collection("demo")
         .when_is("date", "1992-09-17")
         .get_first()
@@ -377,17 +292,17 @@ fn update_test() {
         data: HashMap::from([
             (String::from("special-item"), String::from("false")),
             (String::from("date"), String::from("1992-09-17")),
-        ])
+        ]),
     };
 
     assert_eq!(result, expected);
 
-    store
+    siena(provider.clone())
         .collection("demo")
         .when_is("date", "1992-09-17")
         .set(Vec::from([("special-item", "true")]));
 
-    let result_again = store
+    let result_again = siena(provider.clone())
         .collection("demo")
         .when_is("date", "1992-09-17")
         .get_first()
@@ -400,7 +315,7 @@ fn update_test() {
         data: HashMap::from([
             (String::from("special-item"), String::from("true")),
             (String::from("date"), String::from("1992-09-17")),
-        ])
+        ]),
     };
 
     assert!(result.eq(&expected) && result_again.eq(&expected_again))
@@ -409,17 +324,20 @@ fn update_test() {
 #[test]
 fn create_test() {
     let root_dir = env::current_dir().unwrap();
+    let local_dir = format!(
+        "{}{}",
+        root_dir.display().to_string().as_str(),
+        "/test_data"
+    );
+    let provider = LocalProvider {
+        directory: local_dir,
+    };
 
-    let store = Siena::default()
-        .set_store(Store::Local {
-            directory: format!("{}{}", root_dir.display().to_string().as_str(), "/test_data")
-        });
-
-    store
+    siena(provider.clone())
         .create("demo2", "test3")
         .set(Vec::from([("title", "Title goes here")]));
 
-    let result = store
+    let result = siena(provider.clone())
         .collection("demo2")
         .get_first()
         .unwrap();
@@ -428,9 +346,7 @@ fn create_test() {
         id: String::from("test3"),
         collection: String::from("demo2"),
         file_name: String::from("test3.yml"),
-        data: HashMap::from([
-            (String::from("title"), String::from("Title goes here"))
-        ])
+        data: HashMap::from([(String::from("title"), String::from("Title goes here"))]),
     };
 
     assert_eq!(result, expected);