+++ src/utils.rs
@@ -1,3 +1,3 @@
-pub fn string_ends_with_any(s: &str, suffixes: Vec<&str>) -> bool {
+pub fn str_ends_with_any(s: &str, suffixes: Vec<&str>) -> bool {
     return suffixes.iter().any(|&suffix| s.ends_with(suffix));
 }